CheckPoint-Update at 5pm 30 Jan 2020
Finish CheckPoint, can be improver. CheckPoint work fine when player hit the checkpoint but if player didn't hit any checkpoint it gonna send player to position (0, 0, 0). Leave one checkpoint at the position where player start to solve this problem or fix the code.
