--- IronlightCode/Assets/ART/AmplifyShaderEditor/Examples/HD SRP Samples.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 42169efe5bd48c84e84a420ecbe15d59
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
