--- IronlightCode/Assets/ART/AmplifyShaderEditor/Examples/LW SRP Samples.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bd1e80e339a89454fa9c86cbfe3cf312
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
