--- IronlightCode/Assets/ART/AmplifyShaderEditor/Examples/Official/CustomRTRain/Custom RT Samples.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: db1da8a20db9fa848aed56038e6f24bf
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
