--- IronlightCode/Assets/ART/AmplifyShaderEditor/Examples/Official/Mosaic Effect PPS/MosaicEffectPPS.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6cbf6a1a2e912ed4aa7be068ce29678f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
