--- IronlightCode/Assets/ART/AmplifyShaderEditor/Examples/Official/SimpleTerrain/ApplyGrassWindToTerrain.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8dd36923ebe7b894b931832823664437
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
