--- IronlightCode/Assets/ART/AmplifyShaderEditor/Examples/Official/Substance/LegacySubstanceExample.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7d554bd258399004ebc159e1a3be0a5d
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
