--- IronlightCode/Assets/ART/AmplifyShaderEditor/Examples/Official/Substance2018/Substance2018.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 904dfc5e49f268e45a66be3bd4412718
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
