--- IronlightCode/Assets/ART/AmplifyShaderEditor/Plugins/EditorResources/Templates/CustomRTTemplates.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ee0a6ddfc9eee8e49b6341e94093b0ff
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
