--- IronlightCode/Assets/ART/AmplifyShaderEditor/Plugins/EditorResources/Templates/PPStackTemplates.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3e347f9df0187294384db87fb18b72c4
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
