--- IronlightCode/Assets/ART/AmplifyShaderEditor/Plugins/EditorResources/Templates/SRP/HDSRPTemplates.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0d27534a43933db4c9399f1ad8692742
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
