--- IronlightCode/Assets/ART/AmplifyShaderEditor/Plugins/EditorResources/Templates/SRP/LWSRPTemplates (Legacy).unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c58d43c1e820a964ba0a31216c2e8f97
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
