--- IronlightCode/Assets/ART/AmplifyShaderEditor/Plugins/EditorResources/Templates/SRP/LWSRPTemplates.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9e0806ed7d2436d4a96f58eca1c3bdaf
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
