--- IronlightCode/Assets/ART/CHARACTERS/OWL/Materials/Materials/TutorialLevel.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8bb8183b128469f43a3d161deb1fd592
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
