--- IronlightCode/Assets/ART/CHARACTERS/PLAYER/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 18813126c3199d344b9b908170063d79
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
