--- IronlightCode/Assets/ART/CHARACTERS/SNAKE/Materials/Skull.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ffb892f1adaf143bda37942c61f5bb42
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
