--- IronlightCode/Assets/ART/CHARACTERS/SQUIRREL/Materials/Skull.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 374b651b2a8b2a94cae8e52bde048ebd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
