--- IronlightCode/Assets/ART/David/Alllium/Allium Textures/Flower.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f78975bbfa5082c4d9de5e99762d8997
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
