--- IronlightCode/Assets/ART/David/Log.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d9ddef9f5588ddc4297f15fa56507a4b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
