--- IronlightCode/Assets/ART/David/Models/Shrine.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7ba7dadec44159a42a110d06d0ed361d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
