--- IronlightCode/Assets/ART/David/Models/Tree/Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3e06420ccfbf563439dd759cbe8e14c3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
