--- IronlightCode/Assets/ART/David/Shaders/Water Textures/New Folder.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e1265e5cf05af0f48875444e068a421e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
