--- IronlightCode/Assets/ART/David/Snake/Snake Textures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3dc193d3be54baf4b9fd061655367e9b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
