--- IronlightCode/Assets/ART/David/Textures/Bush.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3aea5d97cf1196544959801f7ea8200e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
