--- IronlightCode/Assets/MAIN/TutorialLevel.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1d82d9ab2c3e848af813f2f58cd8494b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
