+++ IronlightCode/Assets/TESTING/PlayerTestArea.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 7d8906e5a0df6fe4791e7431775e3921
+guid: f443a329072e1ab49b4756f2a25e5d9a
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
