+++ IronlightCode/Assets/TESTING/TESTBUILD(DO_NOT_TOUCH)/ARTMASTER-2.0-ForTESTBUILD.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 30f12915f5ac82b46b64a950fd333936
-folderAsset: yes
+guid: 1c8203cc0cfc4784a848bef23e7c47b5
 DefaultImporter:
   externalObjects: {}
   userData:
