+++ IronlightCode/Assets/TESTING/Viet/CheckPoint/PlayerMovement.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8736e7754a5176b448e21c5cc46bc486
+guid: 3bd54f0e12e9db4419d3e100235c91fe
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
