--- IronlightCode/Assets/TESTING/Viet/CheckPointTesting.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 032552a6ea21dad40b4cf976938cd069
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
