--- IronlightCode/Assets/TESTING/Viet/CheckPointTesting/CheckPointType1.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d9cea9023a5d46440b09ac44a24a2225
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
