--- IronlightCode/Assets/TESTING/Viet/CheckPointTesting/CheckPointType1/CheckPoint.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9ca95c4b3332cf04eb6d842e9faf883f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
