--- IronlightCode/Assets/TESTING/Viet/CheckPointTesting/CheckPointType1/CheckPointTesting.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 63f767993e50fec43a445937cc9166b3
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
